Rename mnist directory
